Refatorando o model e Crud de Cores
