+++ .env
@@ -1,0 +1,6 @@
+<?php
+
+const HOST = 'localhost';
+const DBNAME = 'Sufuture';
+const USER = 'root'; //
+const PASS = '';
